Closes #5754: Refactor PermissionsAdapter into feature addons
